+++ lib/packages/wikibase/data-model/src/Entity/Property.php
@@ -141,4 +141,14 @@ class Property extends Entity {
 		return $this->fingerprint->isEmpty();
 	}
 
+	/**
+	 * Removes all content from the Property.
+	 * The id and the type are not part of the content.
+	 *
+	 * @since 0.1
+	 */
+	public function clear() {
+		$this->fingerprint = Fingerprint::newEmpty();
+	}
+
 }